requirements.txt
botocore==1.24.30
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
httplib2==0.20.4
idna==3.3
jmespath==1.0.0
pycparser==2.21
PyJWT==2.3.0
pynamodb==5.2.1
pyparsing==3.0.7
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
s3transfer==0.5.2
six==1.16.0
twilio==7.8.0
unittest2py3k==0.5.1
urllib3==1.26.9
Werkzeug==2.1.0
